Create implementation_iter.cpp
You need to write function named FindSmallestIndexRange that takes vector (of ints) and two integers (denoting a range of indexes). This function should return an iterator pointer at the smallest element in the vector within the range denoted (inclusive). If multiple elements tie for the smallest, return the first one. Please note that the last argument may denote a position beyond the end of the vector, so make sure you don’t read past the end of the vector. You can assume the range includes at least one element.